crm_django_tenants_migration/requirements.txt
# requirements.txt  ← FULL REPLACEMENT
# Core Django
Django==4.2.16
python-decouple==3.8

# ── Multi-tenancy (PostgreSQL schemas) ───────────────────────────────────────
django-tenants==3.6.1
psycopg2-binary==2.9.9          # PostgreSQL driver (binary = no build deps)
# If you prefer compiled from source: replace with psycopg2==2.9.9

# ── REST API ──────────────────────────────────────────────────────────────────
djangorestframework==3.15.2
django-cors-headers==4.4.0

# ── Auth ─────────────────────────────────────────────────────────────────────
# DRF Token Auth is included with DRF — no extra package needed

# ── Forms & UI (web admin) ────────────────────────────────────────────────────
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0                  # ImageField support

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3

# ── Production extras (uncomment when deploying) ──────────────────────────────
# gunicorn==22.0.0
# whitenoise==6.7.0
# dj-database-url==2.2.0
# django-storages[s3]==1.14.4   # S3 media storage
